% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcalldesc.R
\name{calc.all.desc}
\alias{calc.all.desc}
\title{Retrieve a Java instance of VegajavaR.}
\usage{
calc.all.desc(list.of.smiles, addH = "false", corr.edge.adj = "false",
  aug.corr.ed = "true")
}
\arguments{
\item{list.of.smiles}{vector of smiles}

\item{addH}{if true ad hidrogen to structure}

\item{corr.edge.adj}{if true corrects edge adjacency}

\item{aug.corr.ed}{if true augments and corrects edge adjacency}
}
\value{
matrix with the calculated descriptors
}
\description{
Retrieve a Java instance of VegajavaR, with the training file
specified. Language model is trained before the instance is returned.
The spell corrector is adapted from Peter Norvig's demonstration.
}
