% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegajavaR.R
\name{get_vegajavaR}
\alias{get_vegajavaR}
\title{Retrieve a Java instance of VegajavaR.}
\usage{
get_vegajavaR()
}
\value{
a Java instance of VegajavaR.
}
\description{
Retrieve a Java instance of VegajavaR.
}
